move solc output directory
