+++ .gitignore
@@ -1,6 +1,6 @@
 # Contract ABIs
 build/*
-client/src/build/*
+client/src/contracts/*
 
 # Logs
 logs
